+++ sdk/communication/communication-sms/samples/v1/javascript/optOutCheck.js
@@ -11,48 +11,50 @@ const { SmsClient } = require("@azure/communication-sms");
 require("dotenv").config();
 
 async function main() {
-    console.log("== Opt Out Check ==");
+  console.log("== Opt Out Check ==");
 
-    // You will need to set this environment variable or edit the following values
-    const connectionString =
-        process.env.COMMUNICATION_SAMPLES_CONNECTION_STRING ||
-        "endpoint=https://<resource-name>.communication.azure.com/;<access-key>";
+  // You will need to set this environment variable or edit the following values
+  const connectionString =
+    process.env.COMMUNICATION_SAMPLES_CONNECTION_STRING ||
+    "endpoint=https://<resource-name>.communication.azure.com/;<access-key>";
 
-    // create new client
-    const client = new SmsClient(connectionString);
+  // create new client
+  const client = new SmsClient(connectionString);
 
-    // construct send parameters
-    const from = process.env.FROM_PHONE_NUMBER || process.env.AZURE_PHONE_NUMBER || "<from-phone-number>";
-    let phoneNumbers;
-    if (process.env.TO_PHONE_NUMBERS !== undefined) {
-        phoneNumbers = process.env.TO_PHONE_NUMBERS.split(",");
-    } else if (process.env.AZURE_PHONE_NUMBER !== undefined) {
-        phoneNumbers = [process.env.AZURE_PHONE_NUMBER];
+  // construct send parameters
+  const from =
+    process.env.FROM_PHONE_NUMBER || process.env.AZURE_PHONE_NUMBER || "<from-phone-number>";
+  let phoneNumbers;
+  if (process.env.TO_PHONE_NUMBERS !== undefined) {
+    phoneNumbers = process.env.TO_PHONE_NUMBERS.split(",");
+  } else if (process.env.AZURE_PHONE_NUMBER !== undefined) {
+    phoneNumbers = [process.env.AZURE_PHONE_NUMBER];
+  } else {
+    phoneNumbers = ["<to-phone-number-1>", "<to-phone-number-2>"];
+  }
+
+  // send check opt out request
+  const optOutCheckResults = await client.getOptOutsClient().check(from, phoneNumbers);
+
+  // individual messages can encounter errors during sending
+  // use the "httpStatusCode" property to verify
+  for (const optOutCheckResult of optOutCheckResults) {
+    if (optOutCheckResult.httpStatusCode == 200) {
+      console.log("Success: ", optOutCheckResult);
     } else {
-        phoneNumbers = ["<to-phone-number-1>", "<to-phone-number-2>"];
+      console.error(
+        "Something went wrong when trying to send opt out check request: ",
+        optOutCheckResults,
+      );
     }
+  }
 
-    // send check opt out request
-    const optOutCheckResults = await client.optOuts.check(
-        from,
-        phoneNumbers);
-
-    // individual messages can encounter errors during sending
-    // use the "httpStatusCode" property to verify
-    for (const optOutCheckResult of optOutCheckResults) {
-        if (optOutCheckResult.httpStatusCode == 200) {
-            console.log("Success: ", optOutCheckResult);
-        } else {
-            console.error("Something went wrong when trying to send opt out check request: ", optOutCheckResults);
-        }
-    }
-
-    console.log("== Done: Opt Out Check ==");
+  console.log("== Done: Opt Out Check ==");
 }
 
 main().catch((error) => {
-    console.error("Encountered an error while sending Opt Out Check request: ", error);
-    process.exit(1);
+  console.error("Encountered an error while sending Opt Out Check request: ", error);
+  process.exit(1);
 });
 
 module.exports = { main };